refactor(node): restore owner-crate boundaries for adapter modules
Move the seven audited top-level node adapters to their owners:
consensus/chain/mining own BIP9 context, P2P owns getheaders/getdata,
RPC owns txindex status projection, the txindex worker owns its private
block source, ZMQ and mining own their mempool observers, and bitcoin.conf
parsing lives in the binary input layer. Node keeps composition only.
