+++ crates/node/src/state.rs
@@ -1054,7 +1054,7 @@ fn build_tx_index_open_spec(
 struct TxIndexSpawn {
     spec: crate::txindex_worker::TxIndexOpenSpec,
     generation: crate::txindex_worker::Generation,
-    block_source: crate::NodeBlockSource,
+    block_source: crate::txindex_worker::IndexBlockSource,
     body_source: Arc<dyn BlockBodySource>,
     wake_rx: Receiver<()>,
     recovery_reporter: Arc<crate::recovery_evidence::RecoveryReporter>,
@@ -1082,8 +1082,8 @@ pub struct NodeState {
     tx_index_lifecycle: Option<Arc<arc_swap::ArcSwap<crate::txindex_worker::TxIndexLifecycle>>>,
     /// Stable query adapter for txindex/script-index, constructed before open.
     tx_index_adapter: Option<Arc<crate::txindex_worker::TxIndexQueryAdapter>>,
-    /// Live capability report backing `getcapabilities`.
-    capabilities: Arc<crate::capabilities::NodeCapabilities>,
+    /// Live txindex status for the RPC `getcapabilities` projection.
+    txindex_status: Arc<crate::txindex_worker::TxIndexCapability>,
     prune_service: Option<Arc<dyn PruneService>>,
     zmq_publisher: Arc<dyn crate::ZmqPublisher>,
     active_zmq_notifications: Vec<ZmqNotification>,
@@ -1359,9 +1359,10 @@ impl NodeState {
                     let runtime = Arc::new(crate::txindex_worker::TxIndexRuntime::new(wake_tx));
                     let body_source: Arc<dyn bitcoin_rs_rpc::context::BlockBodySource> =
                         Arc::new(StoredBlockBodySource::new(Arc::clone(&block_body_store)));
-                    let block_source = crate::NodeBlockSource::new(Arc::clone(&blocks))
-                        .with_block_body_source(Arc::clone(&body_source))
-                        .with_block_tree(Arc::clone(&block_tree));
+                    let block_source =
+                        crate::txindex_worker::IndexBlockSource::new(Arc::clone(&blocks))
+                            .with_block_body_source(Arc::clone(&body_source))
+                            .with_block_tree(Arc::clone(&block_tree));
                     let lifecycle: Arc<arc_swap::ArcSwap<crate::txindex_worker::TxIndexLifecycle>> =
                         Arc::new(arc_swap::ArcSwap::from_pointee(
                             crate::txindex_worker::TxIndexLifecycle::Opening,
@@ -1386,12 +1387,10 @@ impl NodeState {
                 }
                 None => (None, None, None, None),
             };
-        let capabilities = Arc::new(crate::capabilities::NodeCapabilities::new(
-            crate::capabilities::CapabilityInputs {
-                tx_lifecycle: tx_index_lifecycle.clone(),
-                tx_runtime: tx_index_runtime.clone(),
-                txindex_enabled: crate::capabilities::txindex_enabled(&config),
-            },
+        let txindex_status = Arc::new(crate::txindex_worker::TxIndexCapability::new(
+            tx_index_lifecycle.clone(),
+            tx_index_runtime.clone(),
+            config.txindex || config.script_index.is_enabled(),
         ));
         let network = Arc::new(RwLock::new(NetworkState::default()));
         let network_active = Arc::new(AtomicBool::new(true));
@@ -1418,26 +1417,22 @@ impl NodeState {
         // `--zmq-pub-sequence` endpoint is configured; the mining
         // generation wake always does.
         let mempool_gateway = {
+            let composite = bitcoin_rs_mempool::CompositeObserver::new();
             let publisher = Arc::clone(&zmq_publisher);
-            let sequence: Option<Arc<dyn bitcoin_rs_mempool::MempoolObserver>> =
-                if publisher.wants_notifications() {
-                    let observer: Arc<dyn bitcoin_rs_mempool::MempoolObserver> = Arc::new(
-                        crate::mempool_observer::MempoolSequenceObserver::new(publisher),
-                    );
-                    Some(observer)
-                } else {
-                    mempool_observer.cloned()
-                };
-            let observer = crate::mempool_observer::NodeMutationObserver::new(
-                sequence,
-                Arc::clone(&mining_generation),
-            );
-            // Intern it: one gateway per pool, so every route that resolves
-            // the pool - including a test attaching its own observer leg -
-            // reaches this instance and its observer slot.
+            if publisher.wants_notifications() {
+                composite.add_leg(
+                    "sequence",
+                    Arc::new(crate::zmq_publisher::MempoolSequenceObserver::new(
+                        publisher,
+                    )),
+                );
+            } else if let Some(observer) = mempool_observer.cloned() {
+                composite.add_leg("test", observer);
+            }
+            composite.add_leg("mining", Arc::clone(&mining_generation) as _);
             bitcoin_rs_mempool::MempoolGateway::shared_with(
                 Arc::clone(&mempool),
-                Arc::new(observer),
+                Arc::new(composite),
             )
         };
         let apply_handles = crate::apply::ApplyHandles {
@@ -1517,7 +1512,7 @@ impl NodeState {
             tx_index_worker: None,
             tx_index_lifecycle,
             tx_index_adapter,
-            capabilities,
+            txindex_status,
             prune_service,
             zmq_publisher,
             active_zmq_notifications,
@@ -1788,10 +1783,10 @@ impl NodeState {
         Ok(())
     }
 
-    /// Returns the live capability report provider for `getcapabilities`.
+    /// Returns the live txindex status source for `getcapabilities`.
     #[must_use]
-    pub fn capability_provider(&self) -> Arc<dyn bitcoin_rs_rpc::context::CapabilityProvider> {
-        self.capabilities.clone()
+    pub fn txindex_status(&self) -> Arc<dyn bitcoin_rs_rpc::capabilities::TxIndexStatusSource> {
+        self.txindex_status.clone()
     }
 
     /// Returns the manual pruning service when pruning is enabled.
